git-svn-id: https://svn.apache.org/repos/asf/incubator/lucene.net/trunk@1294875 13f79535-47bb-0310-9956-ffa450edef68

+++ src/core/Search/DefaultSimilarity.cs
@@ -44,7 +44,7 @@ namespace Lucene.Net.Search
 				numTerms = state.GetLength() - state.GetNumOverlap();
 			else
 				numTerms = state.GetLength();
-			return (float) (state.GetBoost() * LengthNorm(field, numTerms));
+			return (state.GetBoost() * LengthNorm(field, numTerms));
 		}
 		
 		/// <summary>Implemented as <c>1/sqrt(numTerms)</c>. </summary>
